Moved bookmark and share buttons to the trail component
