+++ public/app/features/trails/MetricScene.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 
 import {
   SceneObjectState,
@@ -12,14 +12,13 @@ import {
   PanelBuilders,
   sceneGraph,
 } from '@grafana/scenes';
-import { ToolbarButton, Box, Stack, Icon } from '@grafana/ui';
+import { ToolbarButton, Box, Stack } from '@grafana/ui';
 
 import { getAutoQueriesForMetric } from './AutomaticMetricQueries/AutoQueryEngine';
 import { AutoVizPanel } from './AutomaticMetricQueries/AutoVizPanel';
 import { buildBreakdownActionScene } from './BreakdownScene';
 import { MetricSelectScene } from './MetricSelectScene';
 import { SelectMetricAction } from './SelectMetricAction';
-import { getTrailStore } from './TrailStore/TrailStore';
 import {
   ActionViewDefinition,
   getVariablesWithMetricConstant,
@@ -105,13 +104,7 @@ export class MetricActionBar extends SceneObjectBase<MetricActionBarState> {
   public static Component = ({ model }: SceneComponentProps<MetricActionBar>) => {
     const metricScene = sceneGraph.getAncestor(model, MetricScene);
     const trail = getTrailFor(model);
-    const [isBookmarked, setBookmarked] = useState(false);
     const { actionView } = metricScene.useState();
-
-    const onBookmarkTrail = () => {
-      getTrailStore().addBookmark(trail);
-      setBookmarked(!isBookmarked);
-    };
 
     return (
       <Box paddingY={1}>
@@ -127,19 +120,6 @@ export class MetricActionBar extends SceneObjectBase<MetricActionBarState> {
           ))}
           <ToolbarButton variant={'canvas'}>Add to dashboard</ToolbarButton>
           <ToolbarButton variant={'canvas'} icon="compass" tooltip="Open in explore (todo)" disabled />
-          <ToolbarButton
-            variant={'canvas'}
-            icon={
-              isBookmarked ? (
-                <Icon name={'favorite'} type={'mono'} size={'lg'} />
-              ) : (
-                <Icon name={'star'} type={'default'} size={'lg'} />
-              )
-            }
-            tooltip={'Bookmark'}
-            onClick={onBookmarkTrail}
-          />
-          <ToolbarButton variant={'canvas'} icon="share-alt" tooltip="Copy url (todo)" disabled />
           {trail.state.embedded && (
             <ToolbarButton variant={'canvas'} onClick={model.onOpenTrail}>
               Open
